Remove code duplication in run_for_dataset{,_async}
* call `run_for_dataset_async` from `run_for_dataset`, possibly
  wrapping the `JobExecutor` into an `AsyncAdapter
* merge duplicate utility functions `sync_to_async` and `run_blocking`
* move `async_generator`, `async_generator_eager` to
  `libertem.utils.async_utils`
* new: `async_to_sync_generator` which does exactly what the name says
